♻️ refactor: 지부 도메인 데이터를 entities/organization으로 이전, 미사용 ChapterSelector 삭제
[문제]
shared/ui/segment/ChapterSelector.tsx에 UMC 지부명 목록(CHAPTERS)이 하드코딩되어
있었다. "도메인 무관 공용 컴포넌트" 레이어에 조직 도메인 데이터가 박혀 있어 지부
개편 시 shared/ui를 수정해야 하는 구조. 게다가 이 상수가 features/notice의 barrel을
경유해 라우트들로 재수출되는 우회 경로(shared/ui -> features/notice -> routes)로
퍼져 있었고, isChapter 타입 가드는 3개 라우트 파일에 동일 구현으로 중복돼 있었다.

[조사 결과]
착수 전 소비처를 전수 확인해보니 ChapterSelector 컴포넌트(JSX) 자체는 어디에서도
렌더링되지 않았다. 6개 소비 파일 전부 CHAPTERS 상수/Chapter 타입만 쓰고 실제 지부
탭 UI는 SegmentButton으로 렌더링 중. 즉 컴포넌트는 dead code였고 진짜 문제는 도메인 상수의 위치였다.

[고민한 선택지]
1. 컴포넌트를 items를 받는 형태로 일반화해 shared/ui에 유지 - 렌더링되지 않는
   컴포넌트를 일반화하는 것은 무의미해서 기각. dead code는 삭제가 맞다.
2. CHAPTERS를 getAllChapters API 기반으로 완전 대체 - KPT 회고에서 나온 장기
   방향이지만, 이 상수는 라우트 진입 시점의 URL search param 검증과 기본값 결정에
   동기적으로 쓰이고 있어 비동기 API 데이터로 바꾸면 검증 타이밍이 달라지는 동작
   변경이 된다. 이번 브랜치는 "동작 무변경 구조 정리"가 원칙이라 범위 밖으로 판단,
   TODO 주석으로 남김.
3. (채택) CHAPTERS/Chapter/isChapter를 entities/organization/model/chapters.ts로
   이전. 조직 도메인 데이터의 올바른 레이어이고, 기존에 entities/organization이
   지부 API(getAllChapters)를 이미 소유하고 있어 응집도가 맞다. 6개 소비처는
   entities에서 직접 import하도록 교체(FSD 정방향 유지), notice barrel의 재수출
   우회 경로 제거, 3곳에 중복이던 isChapter 가드는 새 모듈의 단일 구현으로 통합.
