+++ src/features/project/management/ui/ProjectManagementPage.tsx
@@ -13,6 +13,7 @@ import {
 } from "@/entities/member/model/identity"
 import { useViewerIdentity } from "@/entities/member/view-mode/useViewerIdentity"
 import { getChaptersWithSchools } from "@/entities/organization/api/organization"
+import { CHAPTERS } from "@/entities/organization/model/chapters"
 import { projectKeys } from "@/features/project/new/api/queryKeys"
 import { useIsMatchingPeriod } from "@/features/project/new/hooks/useIsMatchingPeriod"
 import { useActiveGisu } from "@/shared/hooks/useActiveGisu"
@@ -20,7 +21,6 @@ import { withImageCacheKey } from "@/shared/lib/withImageCacheKey"
 import { EmptyState } from "@/shared/ui/EmptyState"
 import { ProjectManagementSubTitle } from "@/shared/ui/ProjectManagementSubTitle"
 import { SegmentButton } from "@/shared/ui/segment-button/SegmentButton"
-import { CHAPTERS } from "@/shared/ui/segment/ChapterSelector"
 
 import { getManagedProjects } from "../api"
 import { ProjectManagementCard } from "./ProjectManagementCard"
