+++ src/routes/matching/projects/announce/index.tsx
@@ -16,10 +16,13 @@ import {
   getAllChapters,
   getAllGisu,
 } from "@/entities/organization/api/organization"
-import { ensureMe } from "@/features/auth/lib/ensureMe"
 import {
   type Chapter,
   CHAPTERS,
+  isChapter,
+} from "@/entities/organization/model/chapters"
+import { ensureMe } from "@/features/auth/lib/ensureMe"
+import {
   NoticeCardList,
   NoticeDetailContent,
   type NoticeItem,
@@ -51,12 +54,6 @@ type PendingNotice = {
   title: string
   chip?: string
   mode: "publish" | "edit"
-}
-
-function isChapter(value: unknown): value is Chapter {
-  return (
-    typeof value === "string" && CHAPTERS.some((chapter) => chapter === value)
-  )
 }
 
 function parsePage(value: unknown): number {
